move autocrop to crop plugin
