--- packages/plugin-autocrop/README.md
@@ -1,3 +1,0 @@
-# @jimp/plugin-autocrop
-
-autocrop an image.
